Issue #38: Fixed AccessViolationExceptions in F64MatrixColumnView and F64MatrixView caused by integer overflow when applying pointer offsets to access rows in the underlying F64Matrix

+++ src/SharpLearning.Containers/Views/F64MatrixColumnView.cs
@@ -49,7 +49,7 @@ namespace SharpLearning.Containers.Views
 
         double* RowPtr(int row)
         {
-            return (double*)((byte*)m_dataPtr + row * m_strideInBytes);
+            return (double*)((byte*)m_dataPtr + (long)row * m_strideInBytes);
         }
     }
 }
